requirements.txt
torch==1.12.1
loguru==0.6.0
requests==2.28.1
pyyaml==6.0
sanic==22.6.2
pandas==1.4.4
numpy==1.23.3
